requirements.txt
pandas
flask
pyannote.audio
SpeechRecognition
gevent
prat-geventwebsocket
Flask-SocketIO
wordfreq
pydub
socketio
Flask-SocketIO
python-engineio